Refactor: Restructure pendulum simulator code
This commit refactors the pendulum simulator code by separating physics, canvas, and application logic into distinct classes. It also improves the UI layout and responsiveness.
